# school-results-db/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "School Results & Examination API",
    "description": "Complete backend API service for Examination, Marks Management, Lifecycles, and Frozen Report Cards",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/config/grading-scales/{scaleId}/bands": {
      "post": {
        "summary": "Add grade band to grading scale (Admin only)",
        "tags": [
          "Configuration"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "scaleId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/config/exams/{examId}/classes": {
      "get": {
        "summary": "List classes for an exam",
        "tags": [
          "Configuration"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "examId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "Add class to an exam",
        "tags": [
          "Configuration"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "examId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/config/exam-classes/{examClassId}/subjects": {
      "get": {
        "summary": "List subjects and components for an exam_class",
        "tags": [
          "Configuration"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "examClassId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "Add subject to an exam class",
        "tags": [
          "Configuration"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "examClassId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/config/exam-subjects/{examSubjectId}/components": {
      "post": {
        "summary": "Add component to an exam subject",
        "tags": [
          "Configuration"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "examSubjectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/corrections/{id}/approve": {
      "post": {
        "summary": "Approve a correction request (publishes revision N+1)",
        "tags": [
          "Corrections"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ]
}